Working on adding capability of control f, for finding and syntax highlighting
