+++ pyTexed.py
@@ -1,76 +1,234 @@
-import Tkinter
 from Tkinter import *
-from ScrolledText import *
-import tkFileDialog
-import tkMessageBox
+import os, re
 
-root = Tkinter.Tk(className="Just another Text Editor")
-editor = ScrolledText(root, width=100, height=80)
+TITLE = 'pyTexed'
 
-# create a menu & define functions for each menu item
+FILETYPES = [
+    ("Python Files", "*.py"), ("Text files", "*.txt"), ("All files", "*")
+    ]
 
-def open_command():
-        filename = tkFileDialog.askopenfile(parent=root,mode='rb',title='Select a file')
-        if filename != None:
-            text = open(filename).read()
-            editor.delete(1.0, END)
-            editor.insert(END, text)
-            editor.mark_set(INSERT, 1.0)
+class RoomEditor(Text, object):
 
-def save_command():
-    filename = tkFileDialog.asksaveasfile(parent=root, mode='w',title='Enter a Destination',defaultextension=".txt")
-    if filename != None:
+    modified = None
+
+    def __init__(self, master, **options):
+        Text.__init__(self, master, **options)
+
+        self.config(
+            borderwidth=0,
+            font="{Lucida Sans Typewriter} 11",
+            foreground="#cc9900",
+            background="black",
+            insertbackground="green", # cursor
+            selectforeground="white", # selection
+            selectbackground="#008000",
+            wrap=WORD, # use word wrapping
+            undo=True,
+            width=64,
+            )
+
+        self.filename = None # current document
+
+    def _getfilename(self):
+        return self._filename
+
+    def _setfilename(self, filename):
+        self._filename = filename
+        title = os.path.basename(filename or "(new document)")
+        title = title + " - " + TITLE
+        self.winfo_toplevel().title(title)
+
+    filename = property(_getfilename, _setfilename)
+
+    def edit_modified(self, value=None):
+        # Python 2.5's implementation is broken
+        return self.tk.call(self, "edit", "modified", value)
+
+    def load(self, filename):
+        text = open(filename).read()
+        self.delete(1.0, END)
+        self.insert(END, text)
+        self.mark_set(INSERT, 1.0)
+        self.modified = False
+        self.filename = filename
+
+    def save(self, filename=None):
+        if filename is None:
+            filename = self.filename
         f = open(filename, "w")
-        text = editor.get(1.0, END)
+        s = self.get(1.0, END)
         try:
-            # normalize trailing whitespace
-            f.write(text.rstrip())
+            f.write(s.rstrip())
             f.write("\n")
         finally:
             f.close()
+        self.modified = False
+        self.filename = filename
 
-def exit_command():
-    if tkMessageBox.askokcancel("Quit", "Do you really want to quit?"):
-        root.destroy()
-        sys.exit(0)
+    def clear(self):
+        pass
 
-def about_command():
-    label = tkMessageBox.showinfo("About", "Just Another TextPad \n Copyright \n No rights left to reserve")
+class Cancel(Exception):
+    pass
 
+def open_as():
+    from tkFileDialog import askopenfilename
+    f = askopenfilename(parent=root, filetypes=FILETYPES)
+    if not f:
+        raise Cancel
+    try:
+        editor.load(f)
+    except IOError:
+        from tkMessageBox import showwarning
+        showwarning("Open", "Cannot open the file.")
+        raise Cancel
 
-def dummy():
-    print "I am a Dummy Command, I will be removed in the next step"
+def save_as():
+    from tkFileDialog import asksaveasfilename
+    f = asksaveasfilename(parent=root, defaultextension=".txt")
+    if not f:
+        raise Cancel
+    try:
+        editor.save(f)
+    except IOError:
+        from tkMessageBox import showwarning
+        showwarning("Save As", "Cannot save the file.")
+        raise Cancel
 
-def main():
+def save():
+    if editor.filename:
+        try:
+            editor.save(editor.filename)
+        except IOError:
+            from tkMessageBox import showwarning
+            showwarning("Save", "Cannot save the file.")
+            raise Cancel
+    else:
+        save_as()
+
+def save_if_modified():
+    if not editor.modified:
+        return
+    if askyesnocancel(TITLE, "Document modified. Save changes?"):
+        save()
+
+def askyesnocancel(title=None, message=None, **options):
+    import tkMessageBox
+    s = tkMessageBox.Message(
+        title=title, message=message,
+        icon=tkMessageBox.QUESTION,
+        type=tkMessageBox.YESNOCANCEL,
+        **options).show()
+    if isinstance(s, bool):
+        return s
+    if s == "cancel":
+        raise Cancel
+    return s == "yes"
+
+def file_new(event=None):
+    try:
+        save_if_modified()
+        editor.clear()
+    except Cancel:
+        pass
+    return "break" # don't propagate events
+
+def file_open(event=None):
+    try:
+        save_if_modified()
+        open_as()
+    except Cancel:
+        pass
+    return "break"
+
+def file_save(event=None):
+    try:
+        save()
+    except Cancel:
+        pass
+    return "break"
+
+def file_save_as(event=None):
+    try:
+        save_as()
+    except Cancel:
+        pass
+    return "break"
+
+def file_quit(event=None):
+    try:
+        save_if_modified()
+    except Cancel:
+        return
+    root.quit()
+
+def about(event=None):
+    pass
+
+def find(event=None):
+    newline = "\n"
+    data = editor.get("1.0", END)
+    lines = data.split(newline)
+    p = re.compile('tes', re.IGNORECASE)
+    for line in lines:
+        iterator = p.finditer(line)
+        for match in iterator:
+            res = match.span()
+            sys.stdout.write(str(res))
+        sys.stdout.write(newline)
+
+def main(root):
+
     menu = Menu(root)
     root.config(menu=menu)
+
+    # file menu dro down
     filemenu = Menu(menu)
     menu.add_cascade(label="File", menu=filemenu)
-    filemenu.add_command(label="New", command=dummy)
-    filemenu.add_command(label="Open...", command=open_command)
-    filemenu.add_command(label="Save", command=save_command)
+
+    filemenu.add_command(label="New", command=file_new)
+    filemenu.add_command(label="Open", command=file_open)
+    filemenu.add_command(label="Save", command=file_save)
     filemenu.add_separator()
-    filemenu.add_command(label="Exit", command=exit_command)
+    filemenu.add_command(label="Quit", command=file_quit)
     helpmenu = Menu(menu)
     menu.add_cascade(label="Help", menu=helpmenu)
-    helpmenu.add_command(label="About...", command=about_command)
+    helpmenu.add_command(label="About...", command=about)
 
-    # theme changing
-    editor.pack(fill=Y, expand=1)
-    editor.config(borderwidth=0,
-    font="{Lucida Sans Typewriter} 12",
-    foreground="green",
-    background="black",
-    insertbackground="white", # cursor
-    selectforeground="green", # selection
-    selectbackground="#008000",
-    wrap=WORD, # use word wrapping
-    width=64,
-    undo=True, # Tk 8.4
-    )
+    # tools menu drop down
+    toolsmenu = Menu()
+    toolsmenu.add_cascade(label="Tools", menu=filemenu)
+
+    toolsmenu.add_command(label="Find", command=find)
+
+    # editor set up
+    root.config(background="black")
+    root.wm_state("zoomed")
+
+    editor.pack(fill=Y, expand=1, pady=10)
 
     editor.focus_set()
 
-    root.mainloop()
+    # key bindings for short cuts
+    editor.bind("<Control-n>", file_new)
+    editor.bind("<Control-o>", file_open)
+    editor.bind("<Control-s>", file_save)
+    editor.bind("<Control-Shift-S>", file_save_as)
+    editor.bind("<Control-q>", file_quit)
+    editor.bind("<Control-f>", find)
 
-main()
+    root.protocol("WM_DELETE_WINDOW", file_quit) # window close button
+
+    try:
+        editor.load(sys.argv[1])
+    except (IndexError, IOError):
+        pass
+
+    mainloop()
+
+
+root = Tk()
+editor = RoomEditor(root)
+
+main(root)
+sys.exit(0)
